HP Update
Boss health bar added to the Boss Levels, the arena level we agreed upon
and health % showing within the health bar. Bar needs scaling and text
need a bit of touching up. Both the health percent amount and the health
bar are calculated in percentages now so any amount of health we use,
even after in game upgrades will fit the bar and work with the
percentage.